+++ click_extended/decorators/__init__.py
@@ -4,6 +4,8 @@ from click_extended.decorators.check import *
 from click_extended.decorators.check import __all__ as check_all
 from click_extended.decorators.compare import *
 from click_extended.decorators.compare import __all__ as compare_all
+from click_extended.decorators.convert import *
+from click_extended.decorators.convert import __all__ as convert_all
 from click_extended.decorators.load import *
 from click_extended.decorators.load import __all__ as load_all
 from click_extended.decorators.math import *
@@ -18,6 +20,7 @@ from click_extended.decorators.transform import __all__ as transform_all
 __all__ = [
     *check_all,
     *compare_all,
+    *convert_all,
     *load_all,
     *math_all,
     *misc_all,