+++ click_extended/decorators/convert/__init__.py
@@ -1,0 +1,7 @@
+"""Initialization file for the `click_extended.decorators.convert` module."""
+
+from click_extended.decorators.convert.convert_time import convert_time
+
+__all__ = [
+    "convert_time",
+]